// src/lib/enums.ts
export enum EPetGender {
    MALE = "male",
    FEMALE = "female",
}

export enum EPetSize {
    SMALL = "small",
    MEDIUM = "medium",
    LARGE = "large",
}

export enum EPetCategory {
    CAT = "cat",
    DOG = "dog",
}

export enum EPetGenderUI {
    MALE = "Male",
    FEMALE = "Female",
}

export enum EPetSizeUI {
    SMALL = "Small",
    MEDIUM = "Medium",
    LARGE = "Large",
}

export enum EPetCategoryUI {
    CAT = "Cat",
    DOG = "Dog",
}